Promote py3 branch to master

+++ src/iris/bin/sync_targets.py
@@ -5,7 +5,9 @@ from gevent import monkey, sleep, spawn
 monkey.patch_all()  # NOQA
 
 import logging
+import logging.handlers
 import os
+import uuid
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError, DataError
@@ -72,7 +74,7 @@ if pidfile:
 
 
 def normalize_phone_number(num):
-    return format_number(parse(num.decode('utf-8'), 'US'),
+    return format_number(parse(num, 'US'),
                          PhoneNumberFormat.INTERNATIONAL)
 
 
@@ -102,6 +104,17 @@ def prune_target(engine, target_name, target_type):
     else:
         metrics.incr('others_purged')
 
+    if target_type == 'team':
+        try:
+            # rename team to prevent namespace conflict but preserve the ability to reactivate it in the future
+            new_name = str(uuid.uuid4())
+            engine.execute('''UPDATE `target` SET `active` = FALSE, `name` = %s WHERE `name` = %s AND `type_id` = (SELECT `id` FROM `target_type` WHERE `name` = %s)''', (new_name, target_name, target_type))
+            logger.info('Deleted inactive oncall team %s', target_name)
+        except SQLAlchemyError as e:
+            logger.error('Deleting oncall team %s failed: %s', target_name, e)
+            metrics.incr('sql_errors')
+        return
+
     try:
         engine.execute('''DELETE FROM `target` WHERE `name` = %s AND `type_id` = (SELECT `id` FROM `target_type` WHERE `name` = %s)''', (target_name, target_type))
         logger.info('Deleted inactive target %s', target_name)
@@ -115,11 +128,12 @@ def prune_target(engine, target_name, target_type):
     except SQLAlchemyError as e:
         logger.error('Deleting target %s failed: %s', target_name, e)
         metrics.incr('sql_errors')
+        return
 
 
 def fetch_teams_from_oncall(oncall):
     try:
-        return oncall.get('%steams?fields=name&active=1' % oncall.url).json()
+        return oncall.get('%steams?fields=name&active=1&get_id=1' % oncall.url).json()
     except (ValueError, requests.exceptions.RequestException):
         logger.exception('Failed hitting oncall endpoint to fetch list of team names')
         return []
@@ -163,12 +177,23 @@ def sync_from_oncall(config, engine, purge_old_users=True):
         logger.warning('No users found. Bailing.')
         return
 
-    oncall_team_names = fetch_teams_from_oncall(oncall)
+    # get teams from oncall-api and separate the list of tuples into two lists of name and ids
+    oncall_teams_api_response = fetch_teams_from_oncall(oncall)
+    if not oncall_teams_api_response:
+        logger.warning('No teams found. Bailing.')
+        return
+
+    oncall_team_response = list(zip(*oncall_teams_api_response))
+    oncall_team_names = oncall_team_response[0]
+    oncall_team_ids = oncall_team_response[1]
+    oncall_response_dict_name_key = dict(zip(oncall_team_names, oncall_team_ids))
+    oncall_response_dict_id_key = dict(zip(oncall_team_ids, oncall_team_names))
 
     if not oncall_team_names:
         logger.warning('We do not have a list of team names')
 
     oncall_team_names = set(oncall_team_names)
+    oncall_team_ids = set(oncall_team_ids)
 
     session = sessionmaker(bind=engine)()
 
@@ -187,13 +212,13 @@ def sync_from_oncall(config, engine, purge_old_users=True):
             continue
         contacts[row.mode] = row.destination
 
-    iris_usernames = iris_users.viewkeys()
+    iris_usernames = iris_users.keys()
 
     # users from the oncall endpoints and config files
     metrics.set('users_found', len(oncall_users))
     metrics.set('teams_found', len(oncall_team_names))
     oncall_users.update(get_predefined_users(config))
-    oncall_usernames = oncall_users.viewkeys()
+    oncall_usernames = oncall_users.keys()
 
     # set of users not presently in iris
     users_to_insert = oncall_usernames - iris_usernames
@@ -202,20 +227,20 @@ def sync_from_oncall(config, engine, purge_old_users=True):
     users_to_mark_inactive = iris_usernames - oncall_usernames
 
     # get objects needed for insertion
-    target_types = {name: id for name, id in session.execute('SELECT `name`, `id` FROM `target_type`')}  # 'team' and 'user'
-    modes = {name: id for name, id in session.execute('SELECT `name`, `id` FROM `mode`')}
+    target_types = {name: target_id for name, target_id in session.execute('SELECT `name`, `id` FROM `target_type`')}  # 'team' and 'user'
+    modes = {name: mode_id for name, mode_id in session.execute('SELECT `name`, `id` FROM `mode`')}
     iris_team_names = {name for (name, ) in engine.execute('''SELECT `name` FROM `target` WHERE `type_id` = %s''', target_types['team'])}
-
     target_add_sql = 'INSERT INTO `target` (`name`, `type_id`) VALUES (%s, %s) ON DUPLICATE KEY UPDATE `active` = TRUE'
+    oncall_add_sql = 'INSERT INTO `oncall_team` (`target_id`, `oncall_team_id`) VALUES (%s, %s)'
     user_add_sql = 'INSERT IGNORE INTO `user` (`target_id`) VALUES (%s)'
     target_contact_add_sql = '''INSERT INTO `target_contact` (`target_id`, `mode_id`, `destination`)
                                 VALUES (%s, %s, %s)
                                 ON DUPLICATE KEY UPDATE `destination` = %s'''
 
     # insert users that need to be
-    logger.info('Users to insert (%d)' % len(users_to_insert))
+    logger.info('Users to insert (%d)', len(users_to_insert))
     for username in users_to_insert:
-        logger.info('Inserting %s' % username)
+        logger.info('Inserting %s', username)
         try:
             target_id = engine.execute(target_add_sql, (username, target_types['user'])).lastrowid
             engine.execute(user_add_sql, (target_id, ))
@@ -225,17 +250,17 @@ def sync_from_oncall(config, engine, purge_old_users=True):
             logger.exception('Failed to add user %s' % username)
             continue
         metrics.incr('users_added')
-        for key, value in oncall_users[username].iteritems():
+        for key, value in oncall_users[username].items():
             if value and key in modes:
-                logger.info('%s: %s -> %s' % (username, key, value))
+                logger.info('%s: %s -> %s', username, key, value)
                 engine.execute(target_contact_add_sql, (target_id, modes[key], value, value))
 
     # update users that need to be
-    contact_update_sql = 'UPDATE target_contact SET destination = %s WHERE target_id = (SELECT id FROM target WHERE name = %s) AND mode_id = %s'
-    contact_insert_sql = 'INSERT INTO target_contact (target_id, mode_id, destination) VALUES ((SELECT id FROM target WHERE name = %s), %s, %s)'
-    contact_delete_sql = 'DELETE FROM target_contact WHERE target_id = (SELECT id FROM target WHERE name = %s) AND mode_id = %s'
+    contact_update_sql = 'UPDATE target_contact SET destination = %s WHERE target_id = (SELECT id FROM target WHERE name = %s AND type_id = %s) AND mode_id = %s'
+    contact_insert_sql = 'INSERT INTO target_contact (target_id, mode_id, destination) VALUES ((SELECT id FROM target WHERE name = %s AND type_id = %s), %s, %s)'
+    contact_delete_sql = 'DELETE FROM target_contact WHERE target_id = (SELECT id FROM target WHERE name = %s AND type_id = %s) AND mode_id = %s'
 
-    logger.info('Users to update (%d)' % len(users_to_update))
+    logger.info('Users to update (%d)', len(users_to_update))
     for username in users_to_update:
         try:
             db_contacts = iris_users[username]
@@ -244,44 +269,126 @@ def sync_from_oncall(config, engine, purge_old_users=True):
                 if mode in oncall_contacts and oncall_contacts[mode]:
                     if mode in db_contacts:
                         if oncall_contacts[mode] != db_contacts[mode]:
-                            logger.info('%s: updating %s' % (username, mode))
+                            logger.info('%s: updating %s', username, mode)
                             metrics.incr('user_contacts_updated')
-                            engine.execute(contact_update_sql, (oncall_contacts[mode], username, modes[mode]))
+                            engine.execute(contact_update_sql, (oncall_contacts[mode], username, target_types['user'], modes[mode]))
                     else:
-                        logger.info('%s: adding %s' % (username, mode))
+                        logger.info('%s: adding %s', username, mode)
                         metrics.incr('user_contacts_updated')
-                        engine.execute(contact_insert_sql, (username, modes[mode], oncall_contacts[mode]))
+                        engine.execute(contact_insert_sql, (username, target_types['user'], modes[mode], oncall_contacts[mode]))
                 elif mode in db_contacts:
-                    logger.info('%s: deleting %s' % (username, mode))
+                    logger.info('%s: deleting %s', username, mode)
                     metrics.incr('user_contacts_updated')
-                    engine.execute(contact_delete_sql, (username, modes[mode]))
+                    engine.execute(contact_delete_sql, (username, target_types['user'], modes[mode]))
                 else:
-                    logger.debug('%s: missing %s' % (username, mode))
+                    logger.debug('%s: missing %s', username, mode)
         except SQLAlchemyError as e:
             metrics.incr('users_failed_to_update')
             metrics.incr('sql_errors')
-            logger.exception('Failed to update user %s' % username)
+            logger.exception('Failed to update user %s', username)
             continue
 
-    # sync teams between iris and oncall
-    teams_to_insert = oncall_team_names - iris_team_names
-    teams_to_deactivate = iris_team_names - oncall_team_names
+# sync teams between iris and oncall
 
-    logger.info('Teams to insert (%d)' % len(teams_to_insert))
-    for t in teams_to_insert:
-        logger.info('Inserting %s' % t)
+    # iris_db_oncall_team_ids (team_ids in the oncall_team table)
+    # oncall_team_ids (team_ids from oncall api call)
+    # oncall_team_names (names from oncall api call)
+    # oncall_response_dict_name_key (key value pairs of oncall team names and ids from api call)
+    # oncall_response_dict_id_key same as above but key value inverted
+    # iris_team_names (names from target table)
+    # iris_target_name_id_dict dictionary of target name -> target_id mappings
+    # iris_db_oncall_team_id_name_dict dictionary of oncall team_id -> oncall name mappings
+
+# get all incoming names that match a target check if that target has an entry in oncall table if not make one
+    iris_target_name_id_dict = {name: target_id for name, target_id in engine.execute('''SELECT `name`, `id` FROM `target` WHERE `type_id` = %s''', target_types['team'])}
+
+    matching_target_names = iris_team_names.intersection(oncall_team_names)
+    if matching_target_names:
+        existing_up_to_date_oncall_teams = {name for (name, ) in session.execute('''SELECT `target`.`name` FROM `target` JOIN `oncall_team` ON `oncall_team`.`target_id` = `target`.`id` WHERE `target`.`name` IN :matching_names''', {'matching_names': tuple(matching_target_names)})}
+        # up to date target names that don't have an entry in the oncall_team table yet
+        matching_target_names_no_oncall_entry = matching_target_names - existing_up_to_date_oncall_teams
+
+        for t in matching_target_names_no_oncall_entry:
+            logger.info('Inserting existing team into oncall_team %s', t)
+            try:
+                engine.execute('''UPDATE `target` SET `active` = TRUE WHERE `id` = %s''', iris_target_name_id_dict[t])
+                engine.execute(oncall_add_sql, (iris_target_name_id_dict[t], oncall_response_dict_name_key[t]))
+            except SQLAlchemyError as e:
+                logger.exception('Error inserting oncall_team %s: %s', t, e)
+                continue
+
+# rename all mismatching target names
+
+    iris_db_oncall_team_id_name_dict = {team_id: name for name, team_id in engine.execute('''SELECT target.name, oncall_team.oncall_team_id FROM `target` JOIN `oncall_team` ON oncall_team.target_id = target.id''')}
+
+    iris_db_oncall_team_ids = {oncall_team_id for (oncall_team_id, ) in engine.execute('''SELECT `oncall_team_id` FROM `oncall_team`''')}
+    matching_oncall_ids = oncall_team_ids.intersection(iris_db_oncall_team_ids)
+
+    name_swaps = {}
+
+    # find teams in the iris database whose names have changed
+    for oncall_id in matching_oncall_ids:
+
+        current_name = iris_db_oncall_team_id_name_dict[oncall_id]
+        new_name = oncall_response_dict_id_key[oncall_id]
+        if current_name != new_name:
+            # handle edge case of teams swapping names
+            if not iris_target_name_id_dict.get(new_name, None):
+                target_id_to_rename = iris_target_name_id_dict[current_name]
+                logger.info('Renaming team %s to %s', current_name, new_name)
+                engine.execute('''UPDATE `target` SET `name` = %s, `active` = TRUE WHERE `id` = %s''', (new_name, target_id_to_rename))
+            else:
+                # there is a team swap so rename to a random name to prevent a violation of unique target name constraint
+                new_name = str(uuid.uuid4())
+                target_id_to_rename = iris_target_name_id_dict[current_name]
+                name_swaps[oncall_id] = target_id_to_rename
+                logger.info('Renaming team %s to %s', current_name, new_name)
+                engine.execute('''UPDATE `target` SET `name` = %s, `active` = TRUE WHERE `id` = %s''', (new_name, target_id_to_rename))
+
+    # go back and rename name_swaps to correct value
+    for oncall_id, target_id_to_rename in name_swaps.items():
+        new_name = oncall_response_dict_id_key[oncall_id]
+        engine.execute('''UPDATE `target` SET `name` = %s, `active` = TRUE WHERE `id` = %s''', (new_name, target_id_to_rename))
+
+
+# create new entries for new teams
+
+    # if the team_id doesn't exist in oncall_team at this point then it is a new team.
+    new_team_ids = oncall_team_ids - iris_db_oncall_team_ids
+    logger.info('Teams to insert (%d)' % len(new_team_ids))
+
+    for team_id in new_team_ids:
+        t = oncall_response_dict_id_key[team_id]
+        new_target_id = None
+
+        # add team to target table
+        logger.info('Inserting %s', t)
         try:
-            target_id = engine.execute(target_add_sql, (t, target_types['team'])).lastrowid
+            new_target_id = engine.execute(target_add_sql, (t, target_types['team'])).lastrowid
             metrics.incr('teams_added')
         except SQLAlchemyError as e:
-            logger.exception('Error inserting team %s: %s' % (t, e))
+            logger.exception('Error inserting team %s: %s', t, e)
             metrics.incr('teams_failed_to_add')
             continue
+
+        # add team to oncall_team table
+        if new_target_id:
+            logger.info('Inserting new team into oncall_team %s', t)
+            try:
+                engine.execute(oncall_add_sql, (new_target_id, team_id))
+            except SQLAlchemyError as e:
+                logger.exception('Error inserting oncall_team %s: %s', t, e)
+                continue
+
     session.commit()
     session.close()
 
     # mark users/teams inactive
     if purge_old_users:
+        # find active teams that don't exist in oncall anymore
+        updated_iris_team_names = {name for (name, ) in engine.execute('''SELECT `name` FROM `target` WHERE `type_id` = %s AND `active` = TRUE''', target_types['team'])}
+        teams_to_deactivate = updated_iris_team_names - oncall_team_names
+
         logger.info('Users to mark inactive (%d)' % len(users_to_mark_inactive))
         for username in users_to_mark_inactive:
             prune_target(engine, username, 'user')
@@ -309,7 +416,16 @@ def get_ldap_lists(l, search_strings, parent_list=None):
                              filterstr=filterstr)
         rtype, rdata, rmsgid, serverctrls = l.result3(msgid, timeout=ldap_timeout, resp_ctrl_classes=known_ldap_resp_ctrls)
 
-        results |= {(data[search_strings['list_cn_field']][0], data[search_strings['list_name_field']][0]) for (dn, data) in rdata}
+        for (dn, data) in rdata:
+            cn_field = data[search_strings['list_cn_field']][0]
+            name_field = data[search_strings['list_name_field']][0]
+
+            if isinstance(cn_field, bytes):
+                cn_field = cn_field.decode('utf-8')
+            if isinstance(name_field, bytes):
+                name_field = name_field.decode('utf-8')
+
+            results |= {(cn_field, name_field)}
 
         pctrls = [c for c in serverctrls
                   if c.controlType == SimplePagedResultsControl.controlType]
@@ -340,7 +456,11 @@ def get_ldap_list_membership(l, search_strings, list_name):
                              )
         rtype, rdata, rmsgid, serverctrls = l.result3(msgid, timeout=ldap_timeout, resp_ctrl_classes=known_ldap_resp_ctrls)
 
-        results |= {data[1].get(search_strings['user_mail_field'], [None])[0] for data in rdata}
+        for data in rdata:
+            member = data[1].get(search_strings['user_mail_field'], [None])[0]
+            if isinstance(member, bytes):
+                member = member.decode('utf-8')
+            results |= {member}
 
         pctrls = [c for c in serverctrls
                   if c.controlType == SimplePagedResultsControl.controlType]
@@ -404,12 +524,16 @@ def batch_remove_ldap_memberships(session, list_id, members):
 
 def sync_ldap_lists(ldap_settings, engine):
     try:
+        if 'cert_path' in ldap_settings:
+            ldap.set_option(ldap.OPT_X_TLS_REQUIRE_CERT, ldap.OPT_X_TLS_ALLOW)
         l = ldap.ldapobject.ReconnectLDAPObject(ldap_settings['connection']['url'])
     except Exception:
         logger.exception('Connecting to ldap to get our mailing lists failed.')
         return
 
     try:
+        if 'cert_path' in ldap_settings:
+            l.set_option(ldap.OPT_X_TLS_CACERTFILE, ldap_settings['cert_path'])
         l.simple_bind_s(*ldap_settings['connection']['bind_args'])
     except Exception:
         logger.exception('binding to ldap to get our mailing lists failed.')
@@ -524,7 +648,7 @@ def sync_ldap_lists(ldap_settings, engine):
                     logger.info('Added %s to list %s', member, list_name)
                 except (IntegrityError, DataError):
                     metrics.incr('ldap_memberships_failed_to_add')
-                    logger.warn('Failed adding %s to %s', member, list_name)
+                    logger.warning('Failed adding %s to %s', member, list_name)
 
                 user_add_count += 1
                 if (ldap_add_pause_interval is not None) and (user_add_count % ldap_add_pause_interval) == 0:
@@ -580,6 +704,14 @@ def main():
         # Do ldap mailing list sync *after* we do the normal sync, to ensure we have the users
         # which will be in ldap already populated.
         if ldap_lists:
+
+            if 'ldap_cert_path' in ldap_lists:
+                ldap_cert_path = ldap_lists['ldap_cert_path']
+                if not os.access(ldap_cert_path, os.R_OK):
+                    logger.error("Failed to read ldap_cert_path certificate")
+                    raise IOError
+                else:
+                    ldap_lists['cert_path'] = ldap_cert_path
             list_run_start = time.time()
             sync_ldap_lists(ldap_lists, engine)
             logger.info('Ldap mailing list sync took %.2f seconds', time.time() - list_run_start)
